requirements.txt
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cssselect==1.1.0
cssutils==2.3.0
idna==2.10
importlib-metadata==4.6.1
lxml==4.6.3
mysql-connector-python==8.0.25
premailer==3.9.0
protobuf==3.17.3
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yagmail==0.14.256
zipp==3.5.0
